+++ CHANGELOG.md
@@ -5,14 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.3.0] - 2023-03-03
 ### Added
 - `all`: Build to CommonJS and ES modules
 - `all`: `tsc-alias` package to use aliases in TypeScript
 - `root`: `yarn rsc` Release Sanity Check script
 - `root`: Publish to NPM GitHub Action
+- `@rarimo/nft-checkout`: Mainnet contract to the chain list
 
 ### Fixed
 - `@rarimo/provider`: Solflare wallet naming [issue](https://github.com/rarimo/js-sdk/issues/12)
+- `@rarimo/nft-checkout`: Estimate swap for input native token
 
 ## [1.2.0] - 2023-02-28
 ### Added
@@ -64,7 +68,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Under the hood changes
 - Initiated repo with `@rarimo/provider` and `@rarimo/nft-checkout` packages
 
-[Unreleased]: https://github.com/rarimo/js-sdk/compare/1.2.0...HEAD
+[Unreleased]: https://github.com/rarimo/js-sdk/compare/1.3.0...HEAD
+[1.3.0]: https://github.com/rarimo/js-sdk/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/rarimo/js-sdk/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/rarimo/js-sdk/compare/1.0.1...1.1.0
 [1.0.1]: https://github.com/rarimo/js-sdk/compare/1.0.0...1.0.1